refactor: replace old get share offer by the new one
