+++ src/features/favorites/components/Favorite.tsx
@@ -8,7 +8,7 @@ import { getBookingButtonProperties } from 'features/favorites/helpers/getBookin
 import { getFavoriteDisplayPrice } from 'features/favorites/helpers/getFavoriteDisplayPrice'
 import { useFavoriteFormattedDate } from 'features/favorites/helpers/useFavoriteFormattedDate'
 import { StepperOrigin } from 'features/navigation/RootNavigator/types'
-import { getShareOffer } from 'features/share/helpers/getShareOfferBest'
+import { getShareOffer } from 'features/share/helpers/getShareOffer'
 import { WebShareModal } from 'features/share/pages/WebShareModalBest'
 import { analytics } from 'libs/analytics'
 import { useDistance } from 'libs/location/hooks/useDistance'
