+++ src/features/share/helpers/getShareOffer.tsx
